test: change profile of a user
We currently observe an error when trying to save the new profile via
`AdUserRecord#setProfileId` on PostgreSQL. But when we try it on H2
database, all things are good.

We think it will be good to introduce some IT tests with testcontainers
to be sure that all things go well.

Part of: #35